+++ tests/test_configuration_model.py
@@ -1,12 +1,12 @@
 """This module contains tests for pydantic model configurations.
 
 """
-
+import pytest
 from sphinx.addnodes import desc_annotation
 from sphinx.testing.util import assert_node
 
 from sphinxcontrib.autodoc_pydantic import PydanticModelDocumenter
-from .compatibility import desc_annotation_directive_prefix
+from .compatibility import desc_annotation_directive_prefix, package_is_missing
 
 KWARGS = dict(documenter=PydanticModelDocumenter.objtype,
               deactivate_all=True)
@@ -79,6 +79,168 @@ def test_autodoc_pydantic_model_show_json_false(autodocument):
         '   :module: target.configuration',
         '',
         '   ModelShowJson.',
+        ''
+    ]
+
+    # explicit global
+    actual = autodocument(
+        options_app={"autodoc_pydantic_model_show_json": False},
+        **kwargs)
+    assert actual == result
+
+    # explicit local
+    actual = autodocument(
+        options_doc={"model-show-json": False},
+        **kwargs)
+    assert actual == result
+
+    # explicit local overwrite global
+    actual = autodocument(
+        options_app={"autodoc_pydantic_model_show_json": True},
+        options_doc={"model-show-json": False},
+        **kwargs)
+    assert actual == result
+
+
+@pytest.mark.skipif(package_is_missing('erdantic'), reason="erdantic missing")
+def test_autodoc_pydantic_model_erdantic_figure_true(autodocument):
+    kwargs = dict(object_path='target.configuration.ModelErdanticFigure',
+                  **KWARGS)
+
+    result = [
+        '',
+        '.. py:pydantic_model:: ModelErdanticFigure',
+        '   :module: target.configuration',
+        '',
+        '   ModelErdanticFigure.',
+        '',
+        '',
+        '   .. raw:: html',
+        '',
+        '      <p><details  class="autodoc_pydantic_collapsable_erd">',
+        '      <summary>Show Entity Relationship Diagram</summary>',
+        '',
+        '   .. graphviz::',
+        '',
+        '      digraph "Entity Relationship Diagram" {',
+        '         graph [fontcolor=gray66,',
+        '            fontsize=9,',
+        '            label="Created by erdantic v0.5.0 <https://github.com/drivendataorg/erdantic>",',
+        '            nodesep=0.5,', '            rankdir=LR,', '            ranksep=1.5', '         ];',
+        '         node [fontsize=14,', '            label="\\N",', '            shape=plain', '         ];',
+        '         "target.configuration.ModelErdanticFigure"   [label=<<table border="0" cellborder="1" cellspacing="0"><tr><td port="_root" colspan="2"><b>ModelErdanticFigure</b></td></tr><tr><td>field1</td><td port="field1">int</td></tr><tr><td>field2</td><td port="field2">str</td></tr><tr><td>related</td><td port="related">ModelErdanticFigureRelated</td></tr></table>>,',
+        '            tooltip="target.configuration.ModelErdanticFigure&#xA;&#xA;ModelErdanticFigure.&#xA;"];',
+        '         "target.configuration.ModelErdanticFigureRelated"   [label=<<table border="0" cellborder="1" cellspacing="0"><tr><td port="_root" colspan="2"><b>ModelErdanticFigureRelated</b></td></tr><tr><td>field1</td><td port="field1">int</td></tr><tr><td>field2</td><td port="field2">str</td></tr></table>>,',
+        '            tooltip="target.configuration.ModelErdanticFigureRelated&#xA;&#xA;ModelErdanticFigureRelated.&#xA;"];',
+        '         "target.configuration.ModelErdanticFigure":related:e -> "target.configuration.ModelErdanticFigureRelated":_root:w   [arrowhead=noneteetee];',
+        '      }'
+        '',
+        '',
+        '',
+        '',
+        '   .. raw:: html',
+        '',
+        '      </details></p>',
+        '',
+        ''
+    ]
+
+    # explicit global
+    actual = autodocument(
+        options_app={"autodoc_pydantic_model_erdantic_figure": True},
+        **kwargs)
+    assert actual[:17] == result[:17]
+    assert actual[18:] == result[18:]
+
+    # explicit local
+    actual = autodocument(
+        options_doc={"model-erdantic-figure": True},
+        **kwargs)
+    assert actual[:17] == result[:17]
+    assert actual[18:] == result[18:]
+
+    # explicit local overwrite global
+    actual = autodocument(
+        options_app={"autodoc_pydantic_model_erdantic_figure": False},
+        options_doc={"model-erdantic-figure": True},
+        **kwargs)
+    assert actual[:17] == result[:17]
+    assert actual[18:] == result[18:]
+
+
+@pytest.mark.skipif(package_is_missing('erdantic'), reason="erdantic missing")
+def test_autodoc_pydantic_model_erdantic_figure_collapsed_false(autodocument):
+    kwargs = dict(object_path='target.configuration.ModelErdanticFigure',
+                  **KWARGS)
+
+    result = [
+        '',
+        '.. py:pydantic_model:: ModelErdanticFigure',
+        '   :module: target.configuration',
+        '',
+        '   ModelErdanticFigure.',
+        '',
+        '   .. graphviz::',
+        '',
+        '      digraph "Entity Relationship Diagram" {',
+        '         graph [fontcolor=gray66,', '            fontsize=9,',
+        '            label="Created by erdantic v0.5.0 <https://github.com/drivendataorg/erdantic>",',
+        '            nodesep=0.5,',
+        '            rankdir=LR,',
+        '            ranksep=1.5',
+        '         ];',
+        '         node [fontsize=14,',
+        '            label="\\N",',
+        '            shape=plain',
+        '         ];',
+        '         "target.configuration.ModelErdanticFigure"   [label=<<table border="0" cellborder="1" cellspacing="0"><tr><td port="_root" colspan="2"><b>ModelErdanticFigure</b></td></tr><tr><td>field1</td><td port="field1">int</td></tr><tr><td>field2</td><td port="field2">str</td></tr><tr><td>related</td><td port="related">ModelErdanticFigureRelated</td></tr></table>>,',
+        '            tooltip="target.configuration.ModelErdanticFigure&#xA;&#xA;ModelErdanticFigure.&#xA;"];',
+        '         "target.configuration.ModelErdanticFigureRelated"   [label=<<table border="0" cellborder="1" cellspacing="0"><tr><td port="_root" colspan="2"><b>ModelErdanticFigureRelated</b></td></tr><tr><td>field1</td><td port="field1">int</td></tr><tr><td>field2</td><td port="field2">str</td></tr></table>>,',
+        '            tooltip="target.configuration.ModelErdanticFigureRelated&#xA;&#xA;ModelErdanticFigureRelated.&#xA;"];',
+        '         "target.configuration.ModelErdanticFigure":related:e -> "target.configuration.ModelErdanticFigureRelated":_root:w   [arrowhead=noneteetee];',
+        '      }',
+        '',
+        ''
+    ]
+
+    # explicit global
+    actual = autodocument(
+        options_app={
+            "autodoc_pydantic_model_erdantic_figure": True,
+            "autodoc_pydantic_model_erdantic_figure_collapsed": False
+        },
+        **kwargs)
+    print(actual)
+    assert actual[:10] == result[:10]
+    assert actual[11:] == result[11:]
+
+    # explicit local
+    actual = autodocument(
+        options_doc={"model-erdantic-figure": True, "model-erdantic-figure-collapsed": False},
+        **kwargs)
+    assert actual[:10] == result[:10]
+    assert actual[11:] == result[11:]
+
+    # explicit local overwrite global
+    actual = autodocument(
+        options_app={"autodoc_pydantic_model_erdantic_figure_collapsed": True},
+        options_doc={"model-erdantic-figure": True, "model-erdantic-figure-collapsed": False},
+        **kwargs)
+    assert actual[:10] == result[:10]
+    assert actual[11:] == result[11:]
+
+
+@pytest.mark.skipif(package_is_missing('erdantic'), reason="erdantic missing")
+def test_autodoc_pydantic_model_erdantic_figure_false(autodocument):
+    kwargs = dict(object_path='target.configuration.ModelErdanticFigure',
+                  **KWARGS)
+
+    result = [
+        '',
+        '.. py:pydantic_model:: ModelErdanticFigure',
+        '   :module: target.configuration',
+        '',
+        '   ModelErdanticFigure.',
         ''
     ]